refactor: Renamed CardMenu to EllipsisMenut and Meatball Icon to EllipsisIcon
Relates #68
